requirements.txt
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
outcome==1.3.0.post0
packaging==24.1
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.32.3
schedule==1.2.2
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
